0.0.040: World_Viewer
Add local tile functionality for World_Viewer.

It's now possible to center on local tiles, as well as global tiles. It is also possible to scroll through local or global tiles using arrow keys

Add demoMode for World_Viewer

Demo mode will scroll slowly over the map. It will likely be used for the title screen in the future.